memperbaiki delete

+++ delete.php
@@ -1,8 +1,12 @@
 <?php
 include "koneksi.php";
 
-$id = $_GET['id'];
+if (!isset($_GET['id']) || empty($_GET['id'])) {
+    header("Location: index.php");
+    exit;
+}
 
+$id = (int) $_GET['id'];
 $query = "DELETE FROM data_alat WHERE id = $id";
 $hasil = mysqli_query($koneksi, $query);
